# internal/api/handlers/rental_handler.go
package handlers

import (
	"GoRent/internal/domain/rental"
	"GoRent/internal/service"
	"github.com/gin-gonic/gin"
	"github.com/google/uuid"
	"net/http"
)

type RentalHandler struct {
	rentalService service.RentalService
}

func NewRentalHandler(rentalService service.RentalService) *RentalHandler {
	return &RentalHandler{rentalService: rentalService}
}

// CreateRental godoc
// @Summary      Создать аренду
// @Description  Создает новую аренду автомобиля
// @Tags         rentals
// @Security     ApiKeyAuth
// @Accept       json
// @Produce      json
// @Param        request  body  rental.CreateRequest  true  "Данные аренды"
// @Success      201      {object}  rental.Response
// @Failure      400      {object}  map[string]string
// @Failure      403      {object}  map[string]string
// @Router       /rentals [post]
func (h *RentalHandler) CreateRental(c *gin.Context) {
	var input rental.CreateRequest
	if err := c.ShouldBindJSON(&input); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "invalid input data"})
		return
	}

	userID, exists := c.Get("userID")
	if !exists {
		c.JSON(http.StatusUnauthorized, gin.H{"error": "user not authenticated"})
		return
	}

	userUUID, err := uuid.Parse(userID.(string))
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "invalid user ID"})
		return
	}

	createdRental, err := h.rentalService.CreateRental(c.Request.Context(), input, userUUID)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}

	c.JSON(http.StatusCreated, createdRental)
}

// GetUserRentals godoc
// @Summary      Получить аренды пользователя
// @Tags         rentals
// @Security     ApiKeyAuth
// @Produce      json
// @Success      200  {array}  rental.Response
// @Failure      401  {object}  map[string]string
// @Router       /rentals [get]
func (h *RentalHandler) GetUserRentals(c *gin.Context) {
	userID, exists := c.Get("userID")
	if !exists {
		c.JSON(http.StatusUnauthorized, gin.H{"error": "user not authenticated"})
		return
	}

	userUUID, err := uuid.Parse(userID.(string))
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "invalid user ID"})
		return
	}

	rentals, err := h.rentalService.GetUserRentals(c.Request.Context(), userUUID)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}

	c.JSON(http.StatusOK, rentals)
}

// GetRental godoc
// @Summary      Получить аренду по ID
// @Tags         rentals
// @Security     ApiKeyAuth
// @Produce      json
// @Param        id  path  string  true  "ID аренды"
// @Success      200  {object}  rental.Response
// @Failure      400  {object}  map[string]string
// @Failure      403  {object}  map[string]string
// @Failure      404  {object}  map[string]string
// @Router       /rentals/{id} [get]
func (h *RentalHandler) GetRental(c *gin.Context) {
	rentalID, err := uuid.Parse(c.Param("id"))
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "invalid rental ID"})
		return
	}

	userID, exists := c.Get("userID")
	if !exists {
		c.JSON(http.StatusUnauthorized, gin.H{"error": "user not authenticated"})
		return
	}

	userUUID, err := uuid.Parse(userID.(string))
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "invalid user ID"})
		return
	}

	rental, err := h.rentalService.GetRental(c.Request.Context(), rentalID, userUUID)
	if err != nil {
		if err.Error() == "access denied" {
			c.JSON(http.StatusForbidden, gin.H{"error": "access denied"})
			return
		}
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}

	c.JSON(http.StatusOK, rental)
}

// CancelRental godoc
// @Summary      Отменить аренду
// @Description  Отменяет аренду (только свою)
// @Tags         rentals
// @Security     ApiKeyAuth
// @Param        id  path  string  true  "ID аренды"
// @Success      204
// @Failure      403  {object}  map[string]string
// @Failure      404  {object}  map[string]string
// @Router       /rentals/{id}/cancel [put]
func (h *RentalHandler) CancelRental(c *gin.Context) {
	rentalID, err := uuid.Parse(c.Param("id"))
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "invalid rental ID"})
		return
	}

	userID, exists := c.Get("userID")
	if !exists {
		c.JSON(http.StatusUnauthorized, gin.H{"error": "user not authenticated"})
		return
	}

	userUUID, err := uuid.Parse(userID.(string))
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "invalid user ID"})
		return
	}

	if err := h.rentalService.CancelRental(c.Request.Context(), rentalID, userUUID); err != nil {
		if err.Error() == "access denied" {
			c.JSON(http.StatusForbidden, gin.H{"error": "access denied"})
			return
		}
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}

	c.Status(http.StatusNoContent)
}

// ApproveRental godoc
// @Summary      Подтвердить аренду
// @Description  Доступно только manager и admin
// @Tags         rentals
// @Security     ApiKeyAuth
// @Param        id  path  string  true  "ID аренды"
// @Success      200
// @Failure      400  {object}  map[string]string
// @Failure      403  {object}  map[string]string
// @Router       /rentals/{id}/approve [put]
func (h *RentalHandler) ApproveRental(c *gin.Context) {
	rentalID, err := uuid.Parse(c.Param("id"))
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "invalid rental ID"})
		return
	}

	if err := h.rentalService.ApproveRental(c.Request.Context(), rentalID); err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}

	c.Status(http.StatusOK)
}